+++ tests/testthat/test-gentlg.R
@@ -1,3 +1,5 @@
+df <- data.frame(label = c("boy", "girl"), name = c("Bob", "Lily"), age = c(12, 15))
+
 test_that("custom alignments work", {
   df <- data.frame(label = c("boy", "girl"), name = c("Bob", "Lily"), age = c(12, 15))
 
@@ -45,7 +47,6 @@ test_that("custom alignments work", {
 
 test_that("replace_lead_whitespaces_ind() is replacing whitespaces", {
   # case when there are 2 leading whitespaces, should insert 180 twips
-  df <- data.frame(label = c("boy", "girl"), name = c("Bob", "Lily"), age = c(12, 15))
   expect_no_error(res <- gentlg(huxme = df,
          print.hux = FALSE,
          colheader = c("  Gender", "Name", "Age")))
@@ -58,4 +59,86 @@ test_that("replace_lead_whitespaces_ind() is replacing whitespaces", {
                                 colheader = c("Gender", "Name", "Age")))
   res <- as.character(res[[1]][2, 1])
   expect_equal(res, "\\keepn\\trhdr Gender")
+})
+
+test_that("gentlg() sets the right colwidths when passing a combination of vectors and numeric values", {
+  wcol <- list(c(0.5, 0.3, 0.2), c(0.25))
+  expect_no_error(hux_tables <- gentlg(
+    huxme = list(df, df),
+    wcol = wcol,
+    print.hux = FALSE
+  ))
+
+  for (i in seq_along(wcol)) {
+    ht <- hux_tables[[i]]
+    num_cols <- ncol(ht)
+    actual_colwidths <- as.numeric(huxtable::col_width(ht))
+    if (length(wcol[[i]]) == 1) {
+      expected_colwidths <- c(wcol[[i]], rep((1 - wcol[[i]]) / (num_cols - 1), num_cols - 1))
+    } else {
+      expected_colwidths <- wcol[[i]]
+    }
+    expect_equal(actual_colwidths, expected_colwidths)
+  }
+})
+
+test_that("gentlg() sets the right colwidths when passing all colwidths explicitly", {
+  wcol <- list(c(0.5, 0.3, 0.2), c(0.25, 0.5, 0.25))
+  expect_no_error(hux_tables <- gentlg(
+    huxme = list(df, df),
+    wcol = wcol,
+    print.hux = FALSE
+  ))
+
+  for (i in seq_along(wcol)) {
+    ht <- hux_tables[[i]]
+    num_cols <- ncol(ht)
+    expected_colwidths <- wcol[[i]]
+    actual_colwidths <- as.numeric(huxtable::col_width(ht))
+    expect_equal(expected_colwidths, actual_colwidths)
+  }
+})
+
+test_that("gentlg() validates that if hux is a single data.frame, wcol cannot be a list", {
+  wcol <- list(c(0.5), c(0.4, 0.4, 0.2))
+  expect_error(hux_tables <- gentlg(
+    huxme = df,
+    wcol = wcol,
+    print.hux = FALSE
+  ),
+  "\\'wcol\\' appears to be"
+  )
+})
+
+test_that("gentlg() validates each element in wcol has the correct length", {
+  wcol <- list(c(0.5, 0.3), c(0.4, 0.4, 0.2))
+  expect_error(hux_tables <- gentlg(
+    huxme = list(df, df),
+    wcol = wcol,
+    print.hux = FALSE
+  ),
+  "wcol\\'s length must be 1 or the length of final output"
+  )
+})
+
+test_that("gentlg() validates that the sum of colwidths is 1", {
+  wcol <- list(c(0.5, 0.2, 0.5), c(0.4, 0.4, 0.2))
+  expect_error(hux_tables <- gentlg(
+    huxme = list(df, df),
+    wcol = wcol,
+    print.hux = FALSE
+  ),
+  "wcol not defined properly"
+  )
+})
+
+test_that("gentlg() validates hux length equals wcol length if wcol is a list", {
+  wcol <- list(c(0.5), c(0.4, 0.4, 0.2))
+  expect_error(hux_tables <- gentlg(
+    huxme = list(df, df, df),
+    wcol = wcol,
+    print.hux = FALSE
+  ),
+  "Arguments \\'wcol\\' and \\'huxme\\' must have the same length."
+  )
 })