+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Exported functions `get_file_name()` and `insert_empty_rows()`.
 - Updated `gentlg_single()` to replace, in the first column of tables/listings headers, every leading whitespace with 90 twips (0.0625 inches) left-indentation RTF markup.
+- `gentlg()` argument `wcol` now allows for a list of column width vectors when `huxme` is a list of tables.
 
 # tidytlg 0.11.0